refactor(logic): split turn and diplomacy resolvers (refs #1519)
Phase 1: Extract turn phases under src/turn/phases/, add TurnResolverConfig
and immutable TurnPipelineState-driven pipeline in turn_phase_runner.
Move order acceptance and economy preview helpers to dedicated modules;
document seed mix/LCG constants.

Phase 2: Split diplomacy_resolver into part files (overture, alliance, war,
intervention/CTA, subsidies/relations).

Add test covering resolveTurnForGameWithConfig parity with resolveTurnForGame.

Made-with: Cursor
