+++ packages/colonizethis_logic/lib/src/turn/phases/consumption_phase.dart
@@ -1,0 +1,70 @@
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../../economy/economy_consumption.dart';
+import '../../world/unit_lookup.dart';
+import '../turn_pipeline_state.dart';
+
+/// Consumption phase; returns new pipeline state with updated feeding maps.
+TurnPipelineState runConsumptionPipelinePhase(TurnPipelineState acc) {
+  final game = acc.game;
+  final landFeeding = Map<String, double>.from(
+    acc.landFeedingCoverageByPlayerId,
+  );
+  final navalFeeding = Map<String, double>.from(
+    acc.navalFeedingCoverageByPlayerId,
+  );
+  final idleLabour = Map<String, WorkerIdleCounts>.from(
+    acc.idleLabourByPlayerId,
+  );
+
+  final updatedPlayers = <Player>[];
+
+  for (final player in game.players) {
+    final regimentCounts = regimentTypeCountsForPlayer(
+      game.worldState,
+      player.id,
+    );
+    final shipCounts = shipTypeCountsForPlayer(game.worldState, player.id);
+
+    final result = resolveConsumption(
+      stockpile: player.stockpile,
+      workers: player.workerPool,
+      regimentCountsById: regimentCounts,
+      shipCountsById: shipCounts,
+    );
+
+    double landCoverage;
+    if (result.totalRegiments <= 0) {
+      landCoverage = 1.0;
+    } else {
+      landCoverage = result.fullyFedRegiments / result.totalRegiments;
+      if (landCoverage < 0) landCoverage = 0;
+      if (landCoverage > 1) landCoverage = 1;
+    }
+    landFeeding[player.id] = landCoverage;
+
+    double navalCoverage;
+    if (result.totalShips <= 0) {
+      navalCoverage = 1.0;
+    } else {
+      navalCoverage = result.fullyFedShips / result.totalShips;
+      if (navalCoverage < 0) navalCoverage = 0;
+      if (navalCoverage > 1) navalCoverage = 1;
+    }
+    navalFeeding[player.id] = navalCoverage;
+    idleLabour[player.id] = result.idleLabour;
+    updatedPlayers.add(
+      player.copyWith(
+        stockpile: result.stockpile,
+        workerPool: result.workerPool,
+      ),
+    );
+  }
+
+  return acc.copyWith(
+    game: game.copyWith(players: updatedPlayers),
+    landFeedingCoverageByPlayerId: landFeeding,
+    navalFeedingCoverageByPlayerId: navalFeeding,
+    idleLabourByPlayerId: idleLabour,
+  );
+}
