+++ packages/colonizethis_logic/lib/src/turn/phases/extraction_phase.dart
@@ -1,0 +1,83 @@
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../../constants.dart';
+import '../../world/connectivity_resolver.dart';
+import '../../economy/economy_extraction.dart';
+import '../../economy/resource_extractor.dart';
+import '../../economy/sea_transport.dart';
+import '../turn_seed_constants.dart';
+
+/// Extraction phase: connectivity, land/overseas extraction, interception.
+Game runExtractionPhase(
+  Game state,
+  MapTopology topology,
+  Map<String, TileMapResult>? tileMapByRegion,
+  Map<String, Map<CommodityId, int>> extractedByPlayerId,
+) {
+  if (extractedByPlayerId.isNotEmpty) {
+    return applyExtractionForPlayers(state, extractedByPlayerId);
+  }
+  if (tileMapByRegion == null || tileMapByRegion.isEmpty) {
+    return state;
+  }
+  final connectivity = resolveConnectivity(
+    game: state,
+    tileMapByRegion: tileMapByRegion,
+    topology: topology,
+  );
+  final extraction = computeExtraction(
+    game: state,
+    tileMapByRegion: tileMapByRegion,
+    connectivityResult: connectivity,
+    techCapForPlayer: (playerId) {
+      final player = state.playerById(playerId);
+      return extractionCapForUnlocked(player?.techUnlocked);
+    },
+  );
+  var currentState = state;
+  final updatedPlayers = <Player>[];
+  var extractionSeed =
+      (state.globalGameSeed ?? 0) ^
+      (state.worldState.turnState.turnNumber * kTurnResolutionSeedMix);
+  for (final player in state.players) {
+    var stockpile = player.stockpile;
+    final tot = extraction[player.id];
+    if (tot != null) {
+      stockpile = applyExtractionToStockpile(stockpile, tot.land);
+      logExtractionAutoTransportLand(player.id, tot.land);
+      final cargoHolds = cargoHoldsForHomeFleet(state, player.id);
+      var overseasDelivered = allocateOverseasToStockpile(
+        tot.overseas,
+        cargoHolds: cargoHolds,
+      );
+      logExtractionAutoTransportOverseasAllocation(
+        playerId: player.id,
+        cargoHolds: cargoHolds,
+        overseasTotals: tot.overseas,
+        allocatedToStockpile: overseasDelivered,
+      );
+      if (overseasDelivered.isNotEmpty) {
+        extractionSeed =
+            (extractionSeed * kTurnResolutionLcgMultiplier +
+                kTurnResolutionLcgIncrement) &
+            kTurnResolutionLcgMask;
+        final interception = applyTradeInterception(
+          currentState,
+          player.id,
+          overseasDelivered,
+          seed: extractionSeed ^ player.id.hashCode,
+        );
+        overseasDelivered = interception.reducedDelivered;
+        currentState = currentState.copyWith(
+          worldState: currentState.worldState.copyWith(
+            fleets: interception.updatedFleets,
+          ),
+        );
+      }
+      stockpile = applyExtractionToStockpile(stockpile, overseasDelivered);
+    }
+    updatedPlayers.add(player.copyWith(stockpile: stockpile));
+  }
+  return currentState.copyWith(players: updatedPlayers);
+}
