+++ packages/colonizethis_logic/lib/src/turn/phases/minor_regiment_upgrade_phase.dart
@@ -1,0 +1,5 @@
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../../world/minor_military_parity.dart';
+
+Game runMinorRegimentUpgradePhase(Game game) => applyMinorMilitaryParity(game);
